006_image-recognition/requirements.txt
transformers
torch
torchvision
pillow
